// electron-ui/node-metaverse/lib/classes/Region.d.ts
export declare class Region {
    static CopyMatrix16: number[];
    static CosineTable16: number[];
    static DequantizeTable16: number[];
    static setup: boolean;
    static OO_SQRT_2: number;
    regionName: string;
    regionOwner: UUID;
    regionID: UUID;
    regionSizeX: number;
    regionSizeY: number;
    regionHandle: Long;
    xCoordinate: number;
    yCoordinate: number;
    estateID: number;
    parentEstateID: number;
    regionFlags: RegionFlags;
    mapImage: UUID;
    simAccess: number;
    maxAgents: number;
    billableFactor: number;
    objectBonusFactor: number;
    waterHeight: number;
    terrainRaiseLimit: number;
    terrainLowerLimit: number;
    pricePerMeter: number;
    redirectGridX: number;
    redirectGridY: number;
    useEstateSun: boolean;
    sunHour: number;
    productSKU: string;
    productName: string;
    maxAgents32: number;
    hardMaxAgents: number;
    hardMaxObjects: number;
    cacheID: UUID;
    cpuClassID: number;
    cpuRatio: number;
    coloName: string;
    terrainBase0: UUID;
    terrainBase1: UUID;
    terrainBase2: UUID;
    terrainBase3: UUID;
    terrainDetail0: UUID;
    terrainDetail1: UUID;
    terrainDetail2: UUID;
    terrainDetail3: UUID;
    terrainStartHeight00: number;
    terrainStartHeight01: number;
    terrainStartHeight10: number;
    terrainStartHeight11: number;
    terrainHeightRange00: number;
    terrainHeightRange01: number;
    terrainHeightRange10: number;
    terrainHeightRange11: number;
    handshakeComplete: boolean;
    handshakeCompleteEvent: Subject<void>;
    circuit: Circuit;
    objects: IObjectStore;
    caps: Caps;
    comms: Comms;
    clientEvents: ClientEvents;
    clientCommands: ClientCommands;
    options: BotOptionFlags;
    agent: Agent;
    messageSubscription: Subscription;
    parcelPropertiesSubscription: Subscription;
    terrain: number[][];
    tilesReceived: number;
    terrainComplete: boolean;
    terrainCompleteEvent: Subject<void>;
    parcelsComplete: boolean;
    parcelsCompleteEvent: Subject<void>;
    parcelOverlayComplete: boolean;
    parcelOverlayCompleteEvent: Subject<void>;
    parcelOverlay: ILandBlock[];
    parcels: Record<number, Parcel>;
    parcelsByUUID: Record<string, Parcel>;
    parcelMap: number[][];
    textures: {
        cloudTextureID?: UUID;
        sunTextureID?: UUID;
        moonTextureID?: UUID;
    };
    parcelCoordinates: {
        x: number;
        y: number;
    }[];
    environment: RegionEnvironment;
    timeOffset: number;
    resolver: ObjectResolver;
    agents: Map<string, Avatar>;
    private uploadCost;
    private parcelOverlayReceived;
    constructor(agent: Agent, clientEvents: ClientEvents, options: BotOptionFlags);
    static IDCTColumn16(linein: number[], lineout: number[], column: number): void;
    static IDCTLine16(linein: number[], lineout: number[], line: number): void;
    static InitialSetup(): void;
    getUploadCost(): Promise<number>;
    getParcelProperties(x: number, y: number): Promise<Parcel>;
    getParcels(): Promise<Parcel[]>;
    resetParcels(): void;
    waitForParcelOverlay(): Promise<void>;
    waitForParcels(): Promise<void>;
    waitForTerrain(): Promise<void>;
    getTerrainHeightAtPoint(x: number, y: number): number;
    exportXML(): string;
    activateCaps(seedURL: string): void;
    handshake(handshake: RegionHandshakeMessage): Promise<void>;
    shutdown(): void;
    private static doesBitmapContainCoordinate;
    private resolveParcel;
    private fillParcel;
}